+++ src/main/java/project/backend/business/auth/service/AuthService.java
@@ -14,8 +14,8 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.backend.business.auth.service.implement.KakaoLoginManager;
-import project.backend.dao.auth.repository.RefreshTokenRedisRepository;
-import project.backend.dao.user.entity.User;
+import project.backend.repository.auth.RefreshTokenRedisRepository;
+import project.backend.entity.user.User;
 import project.backend.common.auth.token.RefreshToken;
 import project.backend.common.auth.token.TokenProvider;
 import project.backend.common.auth.token.TokenResponse;
