+++ src/main/java/project/backend/business/user/implement/UserReader.java
@@ -4,8 +4,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import project.backend.common.error.CustomException;
 import project.backend.common.error.ErrorCode;
-import project.backend.dao.user.entity.User;
-import project.backend.dao.user.repository.UserRepository;
+import project.backend.entity.user.User;
+import project.backend.repository.user.UserRepository;
 
 import java.util.Optional;
