+++ src/main/java/project/backend/business/user/service/UserService.java
@@ -3,8 +3,8 @@ package project.backend.business.user.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import project.backend.dao.user.entity.User;
-import project.backend.dao.user.repository.UserRepository;
+import project.backend.entity.user.User;
+import project.backend.repository.user.UserRepository;
 
 @Service
 @RequiredArgsConstructor
