+++ src/main/java/project/backend/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package project.backend.dao;
+package project.backend.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
