+++ src/main/java/project/backend/entity/tag/Tag.java
@@ -1,11 +1,11 @@
-package project.backend.dao.tag.entity;
+package project.backend.entity.tag;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import project.backend.dao.BaseEntity;
+import project.backend.entity.BaseEntity;
 
 import java.util.ArrayList;
 import java.util.List;
