+++ src/main/java/project/backend/repository/post/PostRepository.java
@@ -1,10 +1,10 @@
-package project.backend.dao.post.repository;
+package project.backend.repository.post;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import project.backend.dao.post.entity.Post;
-import project.backend.dao.post.entity.PostStatus;
-import project.backend.dao.user.entity.User;
+import project.backend.entity.post.Post;
+import project.backend.entity.post.PostStatus;
+import project.backend.entity.user.User;
 
 import java.util.List;
