+++ src/main/java/project/backend/repository/user/UserRepository.java
@@ -1,9 +1,9 @@
-package project.backend.dao.user.repository;
+package project.backend.repository.user;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import project.backend.dao.user.entity.User;
+import project.backend.entity.user.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
